rent_a_vehicle/requirements.txt
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
backcall==0.2.0
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
constantly==15.1.0
cryptography==39.0.0
daphne==4.0.0
decorator==5.1.1
Django==4.1.5
djangorestframework==3.14.0
executing==1.2.0
flake8==6.0.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipython==8.8.0
jedi==0.18.2
matplotlib-inline==0.1.6
mccabe==0.7.0
msgpack==1.0.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.14.0
pyOpenSSL==23.0.0
pytz==2022.7
redis==4.4.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.8.0
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
wcwidth==0.2.5
zope.interface==5.5.2
